made the threshold for blink deteciton variable
I test we recognized, that the threshold set for blink detiction
set for the PyGaze method was to high to securly detect blinks.
As a fix for this problem I made this threshold a variable, which
can be changed within a constans.py. The default value is set to
150ms in defaults.py, as it was previously set within the code.
So for everyone not caring about this feature ther should be no
difference in handling PyGaze, but everyone else now has the
possibility to change this threshold to what ever works best.