Update libeyelink.py
Updates libeyelink.py to avoid division by zero crash error possible in RMS/noise calibration.
Previously, if participant moved away from eyetracker before RMS recording began (or eye was "lost"), since nothing would be recorded, len(Xvar) and len(Yvar) would be 0 and the program would crash when attempting to divide by 0.
This just adds a check.  If things went smoothly, it proceeds as before.  If nothing was recorded, it displays a message saying so, then returns to the calibration screen for a retry.
Apologies if this isn't quite elegant, but it worked on my end!