EyeTribe fixes (see description)
1) Default file name bug resolved (file is now named 'default.tsv'
rather than 'default.txt.tsv')

2) Data is now stored on disk when `stop_recording` is called, and not
on each sample. This should solve lag issues on some systems.

3) General pupil size is now calculated properly. If both eyes are
available, it will be the average of both eyes. If only one eye is
available, it will be that eye. If no eyes are available, if will be 0.0
(=missing value for EyeTribe).

4) The 'fix' column in the data file now contains Booleans (rather than
being perpetually False).

5) No more trailing commas in dicts sent to the EyeTribe, to comply with
the newest API (version 0.9.77). This solves issue #62 .