Bug fix for calling array.array.tostring in Python 3.9 and above
From Python 3.2, `tostring` was renamed  `tobytes`. Then, from Python 3.9, `tostring` was dropped completely. While the newer function name makes more sense (as it returns a byte representation), it simply does not exist if we're on Python versions prior to 3.2. Hence we now check if the `array.array` instance has the attribute `tobytes` before calling it, and we fall back on `tostring` if `tobytes` doesn't exist. 

This is awkward, but a lot less awkward than my previous solution of creating a child class of array.array with the single purpose of defining `tostring` with a call to `tobytes`.

```import array.array
class stupid_array(array.array):
    def tostring(self):
        return self.tobytes()
```

HOW STUPID WAS THAT SOLUTION?! I only decided against it because it would change the class, and that would break any code that checks whether a variable's type is array.array.

Do you think this level of backwards support is insane, and that we should just use the newer and more sensible `tobytes`? I've seen lab computers that run Python 2.4 on Windows XP. I'm not ready to have the conversation that they should consider upgrading. Also, maybe they have an ancient EyeLink that requires such an old system. *You don't know their life!*